+++ ChangeLog.md
@@ -1,5 +1,10 @@
 # Changelog for zlibVersion
 
+## 0.2.0.0
+
+* Add --quiet and --help command line options
+* Output where.exe information
+
 ## 0.1.0.0
 
 * Initial version